Debugging issues with BPEC
Tested solvers using simple equations with easy-to-calculate solutions. Problem with correlation function array bounds (should be 2) - fixed. Solver in main program seems to return a value in x that is the same as the initial guess; LU decomposition subroutines did not seem to work properly - fixed (using Gauss elimination). There are still issues with numerical precision in the evaluation of the Jacobian.