Fixing algorithmic errors
Fixed wrong sign of h0: the sign of h0 in the meso_approx expressions has to opposite from that of kb*temp*log(obj_approx%part()) in main.

Also, the constant h0 has to be updated incrementally, so that it stores the free energy of the approximation at each iteration of the loop over chemical potentials.

Finally, in meso_approx the partition function and the correlation shouldn't be multiplied by exp(h0/(hb*temp)). The whole point is to avoid the overflow errors created by such large terms. This is done by redefining the reference energy h0 as the chemical potential increases.