Loop over chemical potentials and solution with Powel
Tests indicate a generally well-behaved algorithm, with maximum error in coverage prediction about 1e-12, except for coverages > 0.75, where the error goes up to 1.4e-06 (coverage 0.85) and drops again at even higher coverages close to 1. Inaccuracies in the parameters are found at these high coverages.

This is potentially due to the (very) high values of the partition function sums which lead to precision errors when calculating the right hand side. A solution would be to use the H0 constant to re-reference energy levels, thereby preventing large values of the partition function sums...