Matlab focusing on BPEC and more Fortran output
The energy re-referencing implementation (H0) did not change the results of the Fortran code. Checked convergence in both codes, seems fine. It appears that the solution is indeed insensitive to the correction parameters close to the solution at these high coverages. The Fortran code minimizes the objective function to a value order 1e-10, see e.g.:

 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  Chemical potential:  0.540000000000001
  Coverage:            0.768399950245753
  Number of iterations:           3
  Minimum value:  4.527118774697337E-010
  at point:  0.879519504213329        -2.852742166647364E-002