Enabled reporting of Jacobians as solution progresses
Newton's method still uses numerical Jacobians here, but also computes the analytical ones and compares.

Note that the calculate_residuals subroutine in meso_approx can be called with an optional argument to control whether Jacobian will be calculated or not (default is to calculate it). This saves time, because when we do function evaluations to calculate numerical Jacobians, we don't care about the Jacobian.

Tests confirm that numerical Jacobians are not accurate enough, e.g. it was observed that the numerical Jacobian was singular, whereas the analytical one had a condition number of about 0.06.