Made initialization procedures external using interface statements
The "import approximation" statement in each of the interfaces in module meso_approx, allows the external initialization subroutine to access all data pertinent to the class approximation object. The subroutine then becomes associated to the host module.

All tests pass on Intel and GNU Fortran. This solves the issue of cluttering the module source code with specific initialization subroutines.